third_party/WebKit/LayoutTests/inspector/map-expected.txt

Tests Map class.

Bug 74084

Running: testMap
Dumping maps:
  First map:
    key1 : undefined
    key2 : undefined
    size : 0
  Second map:
    key1 : undefined
    key2 : undefined
    size : 0

Dumping maps:
  First map:
    key1 : 2
    key2 : undefined
    size : 1
  Second map:
    key1 : undefined
    key2 : undefined
    size : 0

Dumping maps:
  First map:
    key1 : 1
    key2 : undefined
    size : 1
  Second map:
    key1 : undefined
    key2 : undefined
    size : 0

Dumping maps:
  First map:
    key1 : 1
    key2 : undefined
    size : 1
  Second map:
    key1 : 1
    key2 : undefined
    size : 1

Dumping maps:
  First map:
    key1 : 1
    key2 : undefined
    size : 1
  Second map:
    key1 : 2
    key2 : undefined
    size : 1

Dumping maps:
  First map:
    key1 : 1
    key2 : 2
    size : 2
  Second map:
    key1 : 2
    key2 : undefined
    size : 1

Dumping maps:
  First map:
    key1 : 1
    key2 : 2
    size : 2
  Second map:
    key1 : 2
    key2 : 1
    size : 2

Dumping maps:
  First map:
    key1 : 1
    key2 : undefined
    size : 1
  Second map:
    key1 : 2
    key2 : 1
    size : 2

Dumping maps:
  First map:
    key1 : 1
    key2 : undefined
    size : 1
  Second map:
    key1 : 2
    key2 : undefined
    size : 1

Dumping maps:
  First map:
    key1 : 1
    key2 : undefined
    size : 1
  Second map:
    key1 : 2
    key2 : undefined
    size : 1

Dumping maps:
  First map:
    key1 : undefined
    key2 : undefined
    size : 0
  Second map:
    key1 : 2
    key2 : undefined
    size : 1

Dumping maps:
  First map:
    key1 : undefined
    key2 : undefined
    size : 0
  Second map:
    key1 : undefined
    key2 : undefined
    size : 0


Running: testSet
Dumping sets:
  First set:
    key1 : false
    key2 : false
    size : 0
  Second set:
    key1 : false
    key2 : false
    size : 0

Dumping sets:
  First set:
    key1 : true
    key2 : false
    size : 1
  Second set:
    key1 : false
    key2 : false
    size : 0

Dumping sets:
  First set:
    key1 : true
    key2 : false
    size : 1
  Second set:
    key1 : false
    key2 : false
    size : 0

Dumping sets:
  First set:
    key1 : true
    key2 : false
    size : 1
  Second set:
    key1 : true
    key2 : false
    size : 1

Dumping sets:
  First set:
    key1 : true
    key2 : false
    size : 1
  Second set:
    key1 : true
    key2 : false
    size : 1

Dumping sets:
  First set:
    key1 : true
    key2 : true
    size : 2
  Second set:
    key1 : true
    key2 : false
    size : 1

Dumping sets:
  First set:
    key1 : true
    key2 : true
    size : 2
  Second set:
    key1 : true
    key2 : true
    size : 2

Dumping sets:
  First set:
    key1 : true
    key2 : false
    size : 1
  Second set:
    key1 : true
    key2 : true
    size : 2

Dumping sets:
  First set:
    key1 : true
    key2 : false
    size : 1
  Second set:
    key1 : true
    key2 : false
    size : 1

Dumping sets:
  First set:
    key1 : true
    key2 : false
    size : 1
  Second set:
    key1 : true
    key2 : false
    size : 1

Dumping sets:
  First set:
    key1 : false
    key2 : false
    size : 0
  Second set:
    key1 : true
    key2 : false
    size : 1

Dumping sets:
  First set:
    key1 : false
    key2 : false
    size : 0
  Second set:
    key1 : false
    key2 : false
    size : 0


Running: testStringMap
Dumping map:
    size: 0
Dumping map:
    foo => 1
    size: 1
Dumping map:
    bar => 2
    foo => 1
    size: 2
Dumping map:
    bar => 2
    foo => 1
    hasOwnProperty => 3
    size: 3
Dumping map:
    __proto__ => 4
    bar => 2
    foo => 1
    hasOwnProperty => 3
    size: 4
Dumping map:
    __proto__ => 4
    bar => 2
    constructor => 5
    foo => 1
    hasOwnProperty => 3
    size: 5
Dumping map:
    0 => 6
    __proto__ => 4
    bar => 2
    constructor => 5
    foo => 1
    hasOwnProperty => 3
    size: 6
Dumping map:
    0 => 6
    __proto__ => 4
    bar => 2
    constructor => 5
    foo => 42
    hasOwnProperty => 3
    size: 6
Dumping map:
    0 => 6
    __proto__ => 4
    bar => 42
    constructor => 5
    foo => 42
    hasOwnProperty => 3
    size: 6
Dumping map:
    0 => 6
    __proto__ => 4
    bar => 42
    constructor => 5
    foo => 42
    hasOwnProperty => 42
    size: 6
Dumping map:
    0 => 6
    __proto__ => 42
    bar => 42
    constructor => 5
    foo => 42
    hasOwnProperty => 42
    size: 6
Dumping map:
    0 => 6
    __proto__ => 42
    bar => 42
    constructor => 42
    foo => 42
    hasOwnProperty => 42
    size: 6
Dumping map:
    0 => 42
    __proto__ => 42
    bar => 42
    constructor => 42
    foo => 42
    hasOwnProperty => 42
    size: 6
Dumping map:
    __proto__ => 42
    bar => 42
    constructor => 42
    foo => 42
    hasOwnProperty => 42
    size: 5
Dumping map:
    __proto__ => 42
    bar => 42
    foo => 42
    hasOwnProperty => 42
    size: 4
Dumping map:
    bar => 42
    foo => 42
    hasOwnProperty => 42
    size: 3
Dumping map:
    bar => 42
    foo => 42
    size: 2
Dumping map:
    foo => 42
    size: 1
Dumping map:
    size: 0
Dumping map:
    size: 0
Dumping map:
    size: 0
Dumping map:
    size: 0
Dumping map:
    size: 0
Dumping map:
    size: 0
Dumping map:
    size: 0
